chore: divided the tour-package into reusable components and added picture for everest base camp.
