--- components/tour-inquiry-form.tsx
@@ -1,96 +1,0 @@
-"use client"
-
-import type React from "react"
-
-import { useState } from "react"
-import { CalendarIcon } from "lucide-react"
-import { format } from "date-fns"
-
-import { Button } from "@/components/ui/button"
-import { Calendar } from "@/components/ui/calendar"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
-import { Textarea } from "@/components/ui/textarea"
-import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover"
-import { cn } from "@/lib/utils"
-
-export function TourInquiryForm() {
-  const [date, setDate] = useState<Date>()
-  const [isSubmitting, setIsSubmitting] = useState(false)
-  const [isSuccess, setIsSuccess] = useState(false)
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault()
-    setIsSubmitting(true)
-
-    // Simulate form submission
-    setTimeout(() => {
-      setIsSubmitting(false)
-      setIsSuccess(true)
-
-      // Reset success message after 3 seconds
-      setTimeout(() => {
-        setIsSuccess(false)
-      }, 3000)
-    }, 1000)
-  }
-
-  return (
-    <div className="bg-white p-6 rounded-lg shadow-md">
-      <h3 className="text-xl font-bold mb-4">Inquire About This Tour</h3>
-      {isSuccess ? (
-        <div className="bg-green-100 text-green-700 p-4 rounded-md mb-4">
-          Thank you for your inquiry! We will contact you shortly.
-        </div>
-      ) : null}
-      <form onSubmit={handleSubmit} className="space-y-4">
-        <div className="grid grid-cols-1 gap-4">
-          <div>
-            <Label htmlFor="name">Your Name</Label>
-            <Input id="name" placeholder="John Doe" required />
-          </div>
-          <div>
-            <Label htmlFor="email">Email Address</Label>
-            <Input id="email" type="email" placeholder="john@example.com" required />
-          </div>
-          <div>
-            <Label htmlFor="phone">Phone Number</Label>
-            <Input id="phone" placeholder="+1 234 567 8900" />
-          </div>
-          <div>
-            <Label>Preferred Travel Date</Label>
-            <Popover>
-              <PopoverTrigger asChild>
-                <Button
-                  variant="outline"
-                  className={cn("w-full justify-start text-left font-normal", !date && "text-muted-foreground")}
-                >
-                  <CalendarIcon className="mr-2 h-4 w-4" />
-                  {date ? format(date, "PPP") : "Select date"}
-                </Button>
-              </PopoverTrigger>
-              <PopoverContent className="w-auto p-0">
-                <Calendar mode="single" selected={date} onSelect={setDate} initialFocus />
-              </PopoverContent>
-            </Popover>
-          </div>
-          <div>
-            <Label htmlFor="people">Number of People</Label>
-            <Input id="people" type="number" min="1" placeholder="2" required />
-          </div>
-          <div>
-            <Label htmlFor="message">Your Message</Label>
-            <Textarea
-              id="message"
-              placeholder="Please provide any specific requirements or questions you have about the tour."
-              rows={4}
-            />
-          </div>
-        </div>
-        <Button type="submit" className="w-full bg-teal-600 hover:bg-teal-700" disabled={isSubmitting}>
-          {isSubmitting ? "Submitting..." : "Send Inquiry"}
-        </Button>
-      </form>
-    </div>
-  )
-}
