+++ components/ui/date-picker-field.tsx
@@ -1,0 +1,33 @@
+import { useState } from "react"
+import { CalendarIcon } from "lucide-react"
+import { format } from "date-fns"
+import { Button } from "@/components/ui/button"
+import { Calendar } from "@/components/ui/calendar"
+import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover"
+import { Label } from "@/components/ui/label"
+import { cn } from "@/lib/utils"
+
+interface DatePickerFieldProps {
+  date: Date | undefined
+  setDate: (date: Date | undefined) => void
+}
+
+export const DatePickerField = ({ date, setDate }: DatePickerFieldProps) => (
+  <div>
+    <Label>Preferred Travel Date</Label>
+    <Popover>
+      <PopoverTrigger asChild>
+        <Button
+          variant="outline"
+          className={cn("w-full justify-start text-left font-normal", !date && "text-muted-foreground")}
+        >
+          <CalendarIcon className="mr-2 h-4 w-4" />
+          {date ? format(date, "PPP") : "Select date"}
+        </Button>
+      </PopoverTrigger>
+      <PopoverContent className="w-auto p-0">
+        <Calendar mode="single" selected={date} onSelect={setDate} initialFocus />
+      </PopoverContent>
+    </Popover>
+  </div>
+)
